+++ app/clubs/consts.py
@@ -17,7 +17,7 @@ INITIAL_CLUB_ROLES: list[ClubRoleType] = [
         "default": True,
         "permissions": [
             "clubs.view_club",
-            "clubs.view_event",
+            "events.view_event",
             "clubs.view_team",
         ],
     },
@@ -27,8 +27,8 @@ INITIAL_CLUB_ROLES: list[ClubRoleType] = [
         "permissions": [
             "clubs.view_club",
             "clubs.change_club",
-            "clubs.view_event",
-            "clubs.change_event",
+            "events.view_event",
+            "events.change_event",
             "clubs.view_team",
             "clubs.change_team",
         ],
