+++ app/clubs/signals.py
@@ -2,31 +2,7 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 
 from clubs.consts import INITIAL_CLUB_ROLES
-from clubs.models import Club, ClubRole, Event, EventAttendanceLink, RecurringEvent
-from clubs.services import ClubService
-
-
-@receiver(post_save, sender=RecurringEvent)
-def on_save_recurring_event(sender, instance: RecurringEvent, created=False, **kwargs):
-    """Automations to run when a recurring event is saved."""
-
-    if not created:
-        # Only proceed if model is being saved for first time
-        return
-
-    ClubService.sync_recurring_event(instance)
-
-
-@receiver(post_save, sender=Event)
-def on_save_event(sender, instance: Event, created=False, **kwargs):
-    """Automations to run when event is saved."""
-
-    if not created:
-        # Only proceed if event is being created
-        return
-
-    link = EventAttendanceLink.objects.create(event=instance, reference="Default")
-    link.generate_qrcode()
+from clubs.models import Club, ClubRole
 
 
 @receiver(post_save, sender=Club)
