+++ app/clubs/tests/test_club_perms.py
@@ -1,7 +1,8 @@
 from clubs.models import ClubRole
 from clubs.services import ClubService
-from clubs.tests.utils import create_test_club, create_test_event, create_test_team
+from clubs.tests.utils import create_test_club, create_test_team
 from core.abstracts.tests import TestsBase
+from events.tests.utils import create_test_event
 from users.tests.utils import create_test_user
 
 
@@ -82,7 +83,7 @@ class ClubScopedPermsTests(TestsBase):
     def test_club_event_perms(self):
         """Event permissions should be scoped to a club."""
 
-        event1 = create_test_event(club=self.club1)
+        event1 = create_test_event(host=self.club1)
         self.assertTrue(self.user.has_perm("clubs.view_event", event1))
         self.assertFalse(self.user.has_perm("clubs.change_event", event1))
 
@@ -91,7 +92,7 @@ class ClubScopedPermsTests(TestsBase):
         self.assertTrue(self.user.has_perm("clubs.change_event", event1))
 
         # Test access to other club's events
-        event2 = create_test_event(club=self.club2)
+        event2 = create_test_event(host=self.club2)
         self.assertFalse(self.user.has_perm("clubs.view_event", event2))
         self.assertFalse(self.user.has_perm("clubs.change_event", event2))
