+++ app/clubs/views.py
@@ -2,14 +2,12 @@
 Club views for API and rendering html pages.
 """
 
-import re
-
 from django.contrib.auth.decorators import login_required
-from django.http import FileResponse, HttpRequest
+from django.http import HttpRequest
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 
-from clubs.models import Club, Event
+from clubs.models import Club
 from clubs.services import ClubService
 
 
@@ -30,39 +28,6 @@ def club_home_view(request: HttpRequest, club_id: int):
     club = get_object_or_404(Club, id=club_id)
 
     return render(request, "clubs/club_home.html", context={"club": club})
-
-
-@login_required()
-def record_attendance_view(request: HttpRequest, club_id: int, event_id: int):
-    """Records a club member attended an event."""
-    event = get_object_or_404(Event, id=event_id)
-    ClubService(event.club).record_event_attendance(request.user, event)
-
-    return redirect("clubs:joinevent_done", club_id=club_id, event_id=event_id)
-
-
-def download_event_calendar(request: HttpRequest, club_id: int, event_id: int):
-    club = get_object_or_404(Club, id=club_id)
-    event = get_object_or_404(Event, id=event_id)
-
-    club_svc = ClubService(club)
-    file = club_svc.get_event_calendar(event)
-
-    club_name = re.sub(r"\s+", "_", club.name)
-    event_name = re.sub(r"\s+", "_", event.name)
-    return FileResponse(
-        file, as_attachment=True, filename=f"{club_name}_{event_name}.ics"
-    )
-
-
-def download_club_calendar(request: HttpRequest, club_id: int):
-    club = get_object_or_404(Club, id=club_id)
-
-    club_svc = ClubService(club)
-    file = club_svc.get_calendar()
-
-    club_name = re.sub(r"\s+", "_", club.name)
-    return FileResponse(file, as_attachment=True, filename=f"{club_name}.ics")
 
 
 @login_required()
