+++ app/events/apps.py
@@ -1,0 +1,11 @@
+from django.apps import AppConfig
+
+
+class EventsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "events"
+
+    def ready(self) -> None:
+        from . import signals  # noqa: F401
+
+        return super().ready()
