+++ app/events/templates/events/attendance_done.html
